# Add tsdb.IntersectTwoPostings