#### What this PR does

The implementation of `index.Intersect` can be seen as too complex for when only two postings iterators are provided.

This is a simpler implementation, which is shown to be faster by the provided benchmark.

```
name                              old time/op    new time/op    delta
Intersect/Few_common_postings-16    87.9ns ± 7%    51.6ns ±14%  -41.31%  (p=0.008 n=5+5)
Intersect/All_common_postings-16    62.6ns ± 3%    31.8ns ± 2%  -49.20%  (p=0.008 n=5+5)
Intersect/LongPostings1-16          51.6ns ± 2%    29.4ns ± 1%  -43.01%  (p=0.008 n=5+5)
Intersect/LongPostings2-16          54.0ns ± 5%    29.8ns ± 2%  -44.68%  (p=0.008 n=5+5)

name                              old alloc/op   new alloc/op   delta
Intersect/Few_common_postings-16     64.0B ± 0%     48.0B ± 0%  -25.00%  (p=0.008 n=5+5)
Intersect/All_common_postings-16     64.0B ± 0%     48.0B ± 0%  -25.00%  (p=0.008 n=5+5)
Intersect/LongPostings1-16           64.0B ± 0%     48.0B ± 0%  -25.00%  (p=0.008 n=5+5)
Intersect/LongPostings2-16           64.0B ± 0%     48.0B ± 0%  -25.00%  (p=0.008 n=5+5)

name                              old allocs/op  new allocs/op  delta
Intersect/Few_common_postings-16      2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.008 n=5+5)
Intersect/All_common_postings-16      2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.008 n=5+5)
Intersect/LongPostings1-16            2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.008 n=5+5)
Intersect/LongPostings2-16            2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.008 n=5+5)
```

#### Which issue(s) this PR fixes or relates to

Some followup on https://github.com/grafana/mimir/pull/3048

#### Checklist

- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`