@colega 2022-09-28
I also benchmarked the `ExpandedPostings` usage, which didn't show a substantial improvement, performing even worse in some scenarios:

```
name                                                                  old time/op  new time/op  delta
BucketIndexReader_ExpandedPostings/n="1"-16                           8.33ms ± 6%  8.53ms ± 7%    ~     (p=0.548 n=5+5)
BucketIndexReader_ExpandedPostings/n="1",j="foo"-16                   26.3ms ± 1%  26.4ms ± 2%    ~     (p=0.690 n=5+5)
BucketIndexReader_ExpandedPostings/j="foo",n="1"-16                   26.6ms ± 0%  26.2ms ± 0%  -1.42%  (p=0.008 n=5+5)
BucketIndexReader_ExpandedPostings/n="1",j!="foo"-16                  22.5ms ± 0%  22.9ms ± 0%  +1.58%  (p=0.008 n=5+5)
BucketIndexReader_ExpandedPostings/i=~".*"-16                          105ms ± 1%   108ms ± 5%  +3.22%  (p=0.008 n=5+5)
BucketIndexReader_ExpandedPostings/i=~".+"-16                          294ms ± 1%   295ms ± 1%    ~     (p=0.690 n=5+5)
BucketIndexReader_ExpandedPostings/i=~""-16                            325ms ± 1%   326ms ± 0%    ~     (p=0.190 n=5+4)
BucketIndexReader_ExpandedPostings/i!=""-16                            294ms ± 1%   294ms ± 2%    ~     (p=0.841 n=5+5)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",j="foo"-16           29.1ms ± 1%  28.4ms ± 2%  -2.68%  (p=0.008 n=5+5)
BucketIndexReader_ExpandedPostings/n="1",i=~".*",i!="2",j="foo"-16    33.1ms ± 1%  32.5ms ± 1%  -2.03%  (p=0.008 n=5+5)
BucketIndexReader_ExpandedPostings/n="1",i!=""-16                      148ms ± 1%   148ms ± 1%    ~     (p=0.841 n=5+5)
BucketIndexReader_ExpandedPostings/n="1",i!="",j="foo"-16              158ms ± 0%   159ms ± 1%    ~     (p=0.151 n=5+5)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",j="foo"-16            158ms ± 1%   159ms ± 2%    ~     (p=0.548 n=5+5)
BucketIndexReader_ExpandedPostings/n="1",i=~"1.+",j="foo"-16          26.1ms ± 2%  26.0ms ± 0%    ~     (p=1.000 n=5+4)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!="2",j="foo"-16     166ms ± 4%   166ms ± 1%    ~     (p=1.000 n=5+5)
BucketIndexReader_ExpandedPostings/n="1",i=~".+",i!~"2.*",j="foo"-16   187ms ± 2%   188ms ± 1%    ~     (p=1.000 n=5+5)
BucketIndexReader_ExpandedPostings/i=~"0xxx|1xxx|2xxx"-16             26.5µs ± 1%  27.0µs ± 2%    ~     (p=0.056 n=5+5)
BucketIndexReader_ExpandedPostings/i=~"(0|1|2)xxx"-16                 26.5µs ± 1%  26.7µs ± 1%    ~     (p=0.151 n=5+5)
BucketIndexReader_ExpandedPostings/i=~"[0-2]xxx"-16                   26.5µs ± 1%  26.5µs ± 0%    ~     (p=0.905 n=5+4)
BucketIndexReader_ExpandedPostings/i!~[0-2]xxx-16                      123ms ± 3%   119ms ± 2%  -3.87%  (p=0.016 n=5+5)
BucketIndexReader_ExpandedPostings/i=~".*",_i!~[0-2]xxx-16             129ms ± 2%   124ms ± 1%  -3.72%  (p=0.008 n=5+5)
BucketIndexReader_ExpandedPostings/p!=""-16                           24.9ms ± 1%  27.2ms ± 3%  +9.41%  (p=0.008 n=5+5)
```

But the most interesting thing is, the `LabelValuesCardinality` performs worse, and it performs worse in scnearios where this intersect isn't even called!

```
name                                                                                                            old time/op    new time/op    delta
Ingester_LabelValuesCardinality/no_matchers,___name___label_with_1_value_all_series-4                              353ns ± 1%     373ns ± 2%  +5.67%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_l_label_with_10k_values,_1_series_each-4                              355ns ± 1%     369ns ± 1%  +3.87%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_10_label_with_1k_values,_10_series_each-4                         357ns ± 1%     372ns ± 2%  +4.22%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_100_label_with_100_values,_100_series_each-4                      355ns ± 1%     370ns ± 1%  +4.15%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_l_label_with_10k_values,_1_series_each-4                         547ns ± 0%     557ns ± 1%  +1.95%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_10_label_with_1k_values,_10_series_each-4                    553ns ± 3%     562ns ± 2%    ~     (p=0.095 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_100_label_with_100_values,_100_series_each-4                 550ns ± 1%     559ns ± 1%  +1.62%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___and_mod_10_matchers,_mod_100_label_with_100_values,_100_series_each-4     764ns ± 1%     772ns ± 1%    ~     (p=0.095 n=5+5)
```

So... I think I'm going to drop this.